Working subtitle converter
Corrects ISO-8859-9 (Turkish) character to utf-8 equivalent.  This allows characters that are not in the English alphabet to be displayed correctly.  There is no error checking to see if the conversion has already been done.  If ran against a file hat has already been converted it will change those characters to something unreadable.